requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
python-telegram-bot==20.7
httpx==2.31.0